thomas with a relatively working throttle
This version has thomas's speed controlled by pressing the left or right arrow buttons. Vic told me to take out thomas Speed ticker, but I wasn't sure that was a good idea, so I undid all of the work we did together, and foolishly forgot that if I changed anything, I wouldn't be able to use the redo command to get back what we did, so now the program is stuck like this.